requirements.txt
torch
torchvision
numpy
pandas
Pillow
matplotlib
segmentation-models-pytorch